[INLONG-9102][Agent] Add file utils
put file-related tool classes into an independent package
